highlights: %llu, not %zu, inside dt_print() -- MinGW has no C99 length modifier
Win64.UCRT64.GNU.skiptest.Debug has been failing on

  coefficient_field.c:1435: unknown conversion type character 'z' in format
                            [-Werror=format=]

dt_print() carries __attribute__((format(printf, 2, 3))), and MinGW maps that to
ms_printf, which does not implement %z. Plain fprintf() gets gnu_printf instead --
src/CMakeLists.txt defines __USE_MINGW_ANSI_STDIO=1 for exactly that reason -- which
is why the %zu in highlights/selftests.c compile and this one does not.

develop/pixelpipe_hb.c already carries the same fix and the same note; this is the
only other %z in a dt_print() call in the tree (checked across every .c, .h and .cc
outside external/, including the ones whose format string sits on its own line).

Verified with the MinGW cross-compiler: a format(printf) function reproduces the CI
error on %zu and accepts %llu with the cast.

+++ src/iop/highlights/coefficient_field.c
@@ -1431,9 +1431,13 @@ void _cf_reconstruct(_hl_region_ctx_t *const ctx)
                           + fabsf(sj2 / tj - sp2 / tp);
     region_worth = _hl_region_worth(benefit);
     ctx->region_worth = region_worth; // every floor site in this region reuses the same decision
+    // %llu and a cast below, not %zu: dt_print() carries __attribute__((format(printf))),
+    // which MinGW maps to ms_printf, and that does not implement the C99 length modifier.
+    // Plain fprintf() gets gnu_printf via __USE_MINGW_ANSI_STDIO and does accept %zu, which
+    // is why the ones in selftests.c are fine. Same reasoning as develop/pixelpipe_hb.c.
     if(getenv("HL_REGION_W"))
-      dt_print(DT_DEBUG_ALWAYS, "[hl regionw] px=%zu benefit=%.6f region_worth=%.4f\n", region_pixels,
-               benefit, region_worth);
+      dt_print(DT_DEBUG_ALWAYS, "[hl regionw] px=%llu benefit=%.6f region_worth=%.4f\n",
+               (unsigned long long)region_pixels, benefit, region_worth);
   }
 
   HL_PFOR()
